+++ CHANGELOG.md
@@ -37,6 +37,7 @@ Versioning](https://semver.org/spec/v2.0.0.html) for `libnopegl`.
 - `ngl_jni_get_env()` to attach and get a JNI environment from any thread
 - `CustomTexture` node which allows implementing custom rendering logic
 - Android binding and viewer
+- Automatic color conversions from one set of primaries to another
 
 ### Fixed
 - Crash when using resizable RTTs with time ranges